% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_roll_up}
\alias{data_roll_up}
\title{Roll Up Contiguous Data}
\usage{
data_roll_up(data)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
\dontrun{data_roll_up(data)}

}
