% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_roll_up}
\alias{data_roll_up}
\title{Roll Up Contiguous Data}
\usage{
data_roll_up(data, start_end_lag = 1)
}
\arguments{
\item{data}{a dataframe with attributes:id, dtstart, dtend}

\item{start_end_lag}{defined lag between the end time and the following start
time in days. Set to 1 (day) as default.}
}
\value{
data is returned in a tibble
}
\description{
Rolls up contiguous data time windows based on a set \code{start_end_log}
}
\examples{
\dontrun{data_roll_up(data)}

}
