% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_data}
\alias{read_data}
\title{Read in Data}
\usage{
read_data(filepath, filetype)
}
\arguments{
\item{filepath}{the filepath to the raw data}

\item{filetype}{a character for the file extension ("csv" or "RDS")}
}
\value{
tibble of data or an error for invalid file extensions
}
\description{
Reads in rds or csv data
}
\examples{

\dontrun{read_data(filepath, filetype)}

}
